remove UserAuth
