+++ app/src/main/java/ch/epfl/sdp/blindwar/ui/SplashScreenActivity.kt
@@ -1,12 +1,15 @@
 package ch.epfl.sdp.blindwar.ui
 
+import android.app.Activity
 import android.content.Intent
 import ch.epfl.sdp.blindwar.R
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
 import ch.epfl.sdp.blindwar.BuildConfig
-import ch.epfl.sdp.blindwar.user.UserAuth
+import ch.epfl.sdp.blindwar.database.UserDatabase
+import ch.epfl.sdp.blindwar.user.AppStatistics
+import ch.epfl.sdp.blindwar.user.User
 import com.firebase.ui.auth.AuthMethodPickerLayout
 import com.firebase.ui.auth.AuthUI
 import com.firebase.ui.auth.ErrorCodes
@@ -25,12 +28,12 @@ class SplashScreenActivity : AppCompatActivity() {
         private const val TAG = "SplashScreen"
     }
 
-    private var userAuth = UserAuth()
+//    private var userAuth = UserAuth()
 
     // See: https://developer.android.com/training/basics/intents/result
     private val signInLauncher = registerForActivityResult(
         FirebaseAuthUIActivityResultContract()
-    ) { res -> startActivity(userAuth.onSignInResult(this, res))}
+    ) { res -> startActivity(onSignInResult(this, res))}
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -40,11 +43,73 @@ class SplashScreenActivity : AppCompatActivity() {
     }
 
     private fun checkCurrentUser() {
-        if (userAuth.isSignedIn()) {
+        if (isSignedIn()) {
             startActivity(Intent(this, MainMenuActivity::class.java))
         } else {
-            signInLauncher.launch(userAuth.createSignInIntent())
+            signInLauncher.launch(createSignInIntent())
         }
+    }
+
+    private fun createSignInIntent(): Intent{
+        val providers = arrayListOf(
+            AuthUI.IdpConfig.EmailBuilder().setRequireName(false).build(),
+            AuthUI.IdpConfig.GoogleBuilder().build())
+//            AuthUI.IdpConfig.AnonymousBuilder().build())
+
+        val customLayout = AuthMethodPickerLayout
+            .Builder(R.layout.activity_login)
+            .setGoogleButtonId(R.id.Btn_google)
+            .setEmailButtonId(R.id.Btn_email)
+            .build()
+
+        // Create and launch sign-in intent
+        return AuthUI.getInstance()
+            .createSignInIntentBuilder()
+            .setIsSmartLockEnabled(!BuildConfig.DEBUG /* credentials */, true /* hints */) // -> for TESTs only
+            .setAvailableProviders(providers)
+            .setLogo(R.drawable.logo) // still used if setAuthMethodPickerLayout(customLayout) ?
+            .setTheme(R.style.Theme_BlindWar) // Set theme
+            .setAuthMethodPickerLayout(customLayout)
+            .build()
+    }
+
+    private fun onSignInResult(activity: Activity, result: FirebaseAuthUIAuthenticationResult): Intent? {
+        val response = result.idpResponse
+        if (result.resultCode == AppCompatActivity.RESULT_OK) {
+            Log.i("lastSignin", "OK")
+            // Successfully signed in
+            val user = FirebaseAuth.getInstance().currentUser // =?= Firebase.auth.currentUser
+            // https://www.tabnine.com/code/java/classes/com.google.firebase.auth.FirebaseAuth
+
+            if( user?.metadata?.lastSignInTimestamp == user?.metadata?.creationTimestamp) {
+                // new user: 1st signIn
+                return Intent(activity, NewUserActivity::class.java)
+            } else {
+                /*
+            - should we update the online database with the local cache here ?
+             */
+                return Intent(activity, MainMenuActivity::class.java)
+            }
+        } else {
+            // Sign in failed. If response is null the user canceled the
+            // sign-in flow using the back button. Otherwise check
+            // response.getError().getErrorCode() and handle the error.
+            // ...
+            // Sign in failed
+            if (response == null) {
+                // User pressed back button
+                return null
+            }
+            if (response.error?.errorCode == ErrorCodes.NO_NETWORK) {
+                return null
+            }
+            Log.e(TAG, "Sign-in error: ", response.error)
+            return null
+        }
+    }
+
+    private fun isSignedIn(): Boolean {
+        return Firebase.auth.currentUser != null
     }
 
 }
